refactor(utxo-lib): move addToPsbt, addChangeOutputToPsbt
Wallet-related code should live in `bitgo/wallet`.

Issue: BG-47824
