+++ modules/utxo-lib/src/bitgo/Unspent.ts
@@ -1,11 +1,7 @@
-import { TxOutput, taproot } from 'bitcoinjs-lib';
+import { TxOutput } from 'bitcoinjs-lib';
 import { Network } from '..';
 import { toOutputScript } from '../address';
-import { createOutputScript2of3, createSpendScriptP2tr, createPaymentP2tr, toXOnlyPublicKey } from './outputScripts';
-import { UtxoPsbt } from './UtxoPsbt';
-import { UtxoTransaction } from './UtxoTransaction';
 import { UtxoTransactionBuilder } from './UtxoTransactionBuilder';
-import { isSegwit, ChainCode, RootWalletKeys, scriptTypeForChain, WalletUnspent, KeyName } from './wallet';
 
 /**
  * Public unspent data in BitGo-specific representation.
@@ -122,132 +118,6 @@ export function addToTransactionBuilder<TNumber extends number | bigint>(
 ): void {
   const { txid, vout, script, value } = toPrevOutput(u, txb.network as Network);
   txb.addInput(txid, vout, sequence, script, value);
-}
-
-/**
- * Add a verifiable change output to the PSBT. The change output and all data
- * needed to verify it from public keys only are added to the PSBT.
- *
- * @param psbt the PSBT to add change output to
- * @param rootWalletKeys keys that will be able to spend the output
- * @param chain chain code to use for deriving scripts (and to determine script
- *              type) chain is an API parameter in the BitGo API, and may be
- *              any valid ChainCode
- * @param index derivation index for the change address
- * @param value value of the change output
- */
-export function addChangeOutputToPsbt(
-  psbt: UtxoPsbt<UtxoTransaction<bigint>>,
-  rootWalletKeys: RootWalletKeys,
-  chain: ChainCode,
-  index: number,
-  value: bigint
-): void {
-  const walletKeys = rootWalletKeys.deriveForChainAndIndex(chain, index);
-  const scriptType = scriptTypeForChain(chain);
-  if (scriptType === 'p2tr') {
-    const payment = createPaymentP2tr(walletKeys.publicKeys);
-    const allLeafHashes = payment.redeems!.map((r) => taproot.hashTapLeaf(r.output!));
-
-    psbt.addOutput({
-      script: payment.output!,
-      value,
-      tapTree: payment.tapTree,
-      tapInternalKey: payment.internalPubkey,
-      tapBip32Derivation: [0, 1, 2].map((idx) => {
-        const pubkey = toXOnlyPublicKey(walletKeys.triple[idx].publicKey);
-        const leafHashes: Buffer[] = [];
-        payment.redeems!.forEach((r, idx) => {
-          if (r.pubkeys!.find((pk) => pk.equals(pubkey))) {
-            leafHashes.push(allLeafHashes[idx]);
-          }
-        });
-        return {
-          leafHashes,
-          pubkey,
-          path: walletKeys.paths[idx],
-          masterFingerprint: rootWalletKeys.triple[idx].fingerprint,
-        };
-      }),
-    });
-  } else {
-    const { scriptPubKey, witnessScript, redeemScript } = createOutputScript2of3(walletKeys.publicKeys, scriptType);
-    psbt.addOutput({
-      script: scriptPubKey,
-      value,
-      bip32Derivation: [0, 1, 2].map((idx) => ({
-        pubkey: walletKeys.triple[idx].publicKey,
-        path: walletKeys.paths[idx],
-        masterFingerprint: rootWalletKeys.triple[idx].fingerprint,
-      })),
-    });
-    const outputIndex = psbt.txOutputs.length - 1;
-    if (witnessScript) {
-      psbt.updateOutput(outputIndex, { witnessScript });
-    }
-    if (redeemScript) {
-      psbt.updateOutput(outputIndex, { redeemScript });
-    }
-  }
-}
-
-export function addToPsbt(
-  psbt: UtxoPsbt<UtxoTransaction<bigint>>,
-  u: WalletUnspent<bigint>,
-  rootWalletKeys: RootWalletKeys,
-  signer: KeyName,
-  cosigner: KeyName,
-  network: Network
-): void {
-  const { txid, vout, script, value } = toPrevOutput(u, network);
-  const walletKeys = rootWalletKeys.deriveForChainAndIndex(u.chain, u.index);
-  const scriptType = scriptTypeForChain(u.chain);
-  psbt.addInput({
-    hash: txid,
-    index: vout,
-    witnessUtxo: {
-      script,
-      value,
-    },
-  });
-  const inputIndex = psbt.inputCount - 1;
-  if (!isSegwit(u.chain)) {
-    if (!isUnspentWithPrevTx(u)) {
-      throw new Error('Error, require previous tx to add to PSBT');
-    }
-    psbt.updateInput(inputIndex, { nonWitnessUtxo: u.prevTx });
-  }
-
-  if (scriptType === 'p2tr') {
-    const { controlBlock, witnessScript, leafVersion, leafHash } = createSpendScriptP2tr(walletKeys.publicKeys, [
-      walletKeys[signer].publicKey,
-      walletKeys[cosigner].publicKey,
-    ]);
-    psbt.updateInput(inputIndex, {
-      tapLeafScript: [{ controlBlock, script: witnessScript, leafVersion }],
-      tapBip32Derivation: [signer, cosigner].map((key) => ({
-        leafHashes: [leafHash],
-        pubkey: walletKeys[key].publicKey.slice(1), // 32-byte x-only
-        path: rootWalletKeys.getDerivationPath(rootWalletKeys[key], u.chain, u.index),
-        masterFingerprint: rootWalletKeys[key].fingerprint,
-      })),
-    });
-  } else {
-    const { witnessScript, redeemScript } = createOutputScript2of3(walletKeys.publicKeys, scriptType);
-    psbt.updateInput(inputIndex, {
-      bip32Derivation: [0, 1, 2].map((idx) => ({
-        pubkey: walletKeys.triple[idx].publicKey,
-        path: walletKeys.paths[idx],
-        masterFingerprint: rootWalletKeys.triple[idx].fingerprint,
-      })),
-    });
-    if (witnessScript) {
-      psbt.updateInput(inputIndex, { witnessScript });
-    }
-    if (redeemScript) {
-      psbt.updateInput(inputIndex, { redeemScript });
-    }
-  }
 }
 
 /**
