+++ modules/utxo-lib/src/bitgo/wallet/Unspent.ts
@@ -1,14 +1,23 @@
+import { taproot } from 'bitcoinjs-lib';
+
 import { Network } from '../..';
 import { UtxoTransactionBuilder } from '../UtxoTransactionBuilder';
-import { createOutputScript2of3, scriptTypeForChain } from '../outputScripts';
+import {
+  createOutputScript2of3,
+  createPaymentP2tr,
+  createSpendScriptP2tr,
+  scriptTypeForChain,
+  toXOnlyPublicKey,
+} from '../outputScripts';
 import { toOutputScript } from '../../address';
 import { signInput2Of3, verifySignatureWithPublicKeys } from '../signature';
 import { WalletUnspentSigner } from './WalletUnspentSigner';
-import { RootWalletKeys } from './WalletKeys';
+import { KeyName, RootWalletKeys } from './WalletKeys';
 import { UtxoTransaction } from '../UtxoTransaction';
 import { Triple } from '../types';
-import { toOutput, UnspentWithPrevTx, Unspent } from '../Unspent';
-import { ChainCode } from './chains';
+import { toOutput, UnspentWithPrevTx, Unspent, isUnspentWithPrevTx, toPrevOutput } from '../Unspent';
+import { ChainCode, isSegwit } from './chains';
+import { UtxoPsbt } from '../UtxoPsbt';
 
 export interface WalletUnspent<TNumber extends number | bigint = number> extends Unspent<TNumber> {
   chain: ChainCode;
@@ -86,4 +95,130 @@ export interface WalletUnspentLegacy<TNumber extends number | bigint = number> e
   redeemScript?: string;
   /** @deprecated - obviated by verifyWithUnspent */
   witnessScript?: string;
+}
+
+/**
+ * Add a verifiable change output to the PSBT. The change output and all data
+ * needed to verify it from public keys only are added to the PSBT.
+ *
+ * @param psbt the PSBT to add change output to
+ * @param rootWalletKeys keys that will be able to spend the output
+ * @param chain chain code to use for deriving scripts (and to determine script
+ *              type) chain is an API parameter in the BitGo API, and may be
+ *              any valid ChainCode
+ * @param index derivation index for the change address
+ * @param value value of the change output
+ */
+export function addChangeOutputToPsbt(
+  psbt: UtxoPsbt<UtxoTransaction<bigint>>,
+  rootWalletKeys: RootWalletKeys,
+  chain: ChainCode,
+  index: number,
+  value: bigint
+): void {
+  const walletKeys = rootWalletKeys.deriveForChainAndIndex(chain, index);
+  const scriptType = scriptTypeForChain(chain);
+  if (scriptType === 'p2tr') {
+    const payment = createPaymentP2tr(walletKeys.publicKeys);
+    const allLeafHashes = payment.redeems!.map((r) => taproot.hashTapLeaf(r.output!));
+
+    psbt.addOutput({
+      script: payment.output!,
+      value,
+      tapTree: payment.tapTree,
+      tapInternalKey: payment.internalPubkey,
+      tapBip32Derivation: [0, 1, 2].map((idx) => {
+        const pubkey = toXOnlyPublicKey(walletKeys.triple[idx].publicKey);
+        const leafHashes: Buffer[] = [];
+        payment.redeems!.forEach((r, idx) => {
+          if (r.pubkeys!.find((pk) => pk.equals(pubkey))) {
+            leafHashes.push(allLeafHashes[idx]);
+          }
+        });
+        return {
+          leafHashes,
+          pubkey,
+          path: walletKeys.paths[idx],
+          masterFingerprint: rootWalletKeys.triple[idx].fingerprint,
+        };
+      }),
+    });
+  } else {
+    const { scriptPubKey, witnessScript, redeemScript } = createOutputScript2of3(walletKeys.publicKeys, scriptType);
+    psbt.addOutput({
+      script: scriptPubKey,
+      value,
+      bip32Derivation: [0, 1, 2].map((idx) => ({
+        pubkey: walletKeys.triple[idx].publicKey,
+        path: walletKeys.paths[idx],
+        masterFingerprint: rootWalletKeys.triple[idx].fingerprint,
+      })),
+    });
+    const outputIndex = psbt.txOutputs.length - 1;
+    if (witnessScript) {
+      psbt.updateOutput(outputIndex, { witnessScript });
+    }
+    if (redeemScript) {
+      psbt.updateOutput(outputIndex, { redeemScript });
+    }
+  }
+}
+
+export function addToPsbt(
+  psbt: UtxoPsbt<UtxoTransaction<bigint>>,
+  u: WalletUnspent<bigint>,
+  rootWalletKeys: RootWalletKeys,
+  signer: KeyName,
+  cosigner: KeyName,
+  network: Network
+): void {
+  const { txid, vout, script, value } = toPrevOutput(u, network);
+  const walletKeys = rootWalletKeys.deriveForChainAndIndex(u.chain, u.index);
+  const scriptType = scriptTypeForChain(u.chain);
+  psbt.addInput({
+    hash: txid,
+    index: vout,
+    witnessUtxo: {
+      script,
+      value,
+    },
+  });
+  const inputIndex = psbt.inputCount - 1;
+  if (!isSegwit(u.chain)) {
+    if (!isUnspentWithPrevTx(u)) {
+      throw new Error('Error, require previous tx to add to PSBT');
+    }
+    psbt.updateInput(inputIndex, { nonWitnessUtxo: u.prevTx });
+  }
+
+  if (scriptType === 'p2tr') {
+    const { controlBlock, witnessScript, leafVersion, leafHash } = createSpendScriptP2tr(walletKeys.publicKeys, [
+      walletKeys[signer].publicKey,
+      walletKeys[cosigner].publicKey,
+    ]);
+    psbt.updateInput(inputIndex, {
+      tapLeafScript: [{ controlBlock, script: witnessScript, leafVersion }],
+      tapBip32Derivation: [signer, cosigner].map((key) => ({
+        leafHashes: [leafHash],
+        pubkey: walletKeys[key].publicKey.slice(1), // 32-byte x-only
+        path: rootWalletKeys.getDerivationPath(rootWalletKeys[key], u.chain, u.index),
+        masterFingerprint: rootWalletKeys[key].fingerprint,
+      })),
+    });
+  } else {
+    const { witnessScript, redeemScript } = createOutputScript2of3(walletKeys.publicKeys, scriptType);
+    psbt.updateInput(inputIndex, {
+      bip32Derivation: [0, 1, 2].map((idx) => ({
+        pubkey: walletKeys.triple[idx].publicKey,
+        path: walletKeys.paths[idx],
+        masterFingerprint: rootWalletKeys.triple[idx].fingerprint,
+      })),
+    });
+    if (witnessScript) {
+      psbt.updateInput(inputIndex, { witnessScript });
+    }
+    if (redeemScript) {
+      psbt.updateInput(inputIndex, { redeemScript });
+    }
+  }
 }
